Update for node tree id provider. Missing a few last minute changes that needed variable casting.
git-svn-id: https://svn.apache.org/repos/asf/incubator/vxquery@1465891 13f79535-47bb-0310-9956-ffa450edef68

+++ trunk/vxquery/vxquery-core/src/main/java/org/apache/vxquery/runtime/functions/node/FnCollectionScalarEvaluatorFactory.java
@@ -63,7 +63,7 @@ public class FnCollectionScalarEvaluatorFactory extends AbstractTaggedValueArgum
         final InputSource in = new InputSource();
         final int partition = ctx.getTaskAttemptId().getTaskId().getPartition();
         // TODO Add the data source scan id.
-        final ITreeNodeIdProvider nodeIdProvider = new TreeNodeIdProvider(partition);
+        final ITreeNodeIdProvider nodeIdProvider = new TreeNodeIdProvider((short) partition);
 
         return new AbstractTaggedValueArgumentScalarEvaluator(args) {
             @Override
